+++ upyog/image/composition.py
@@ -20,6 +20,19 @@ def draw_rule_of_thirds(img: Image.Image, thickness=5, opacity=0.4) -> Image.Ima
     return vis.img
 
 
+def assemble_rows(imgs: ImageCollection, ncol: int) -> list:
+    """Assemble images into rows based on the number of columns."""
+    rows, row = [], []
+    for i, img in enumerate(imgs, start=1):
+        row.append(img)
+        if i % ncol == 0:
+            rows.append(row)
+            row = []
+    if row:
+        rows.append(row)
+    return rows
+
+
 def img_join_horizontal(imgs: ImageCollection) -> Image.Image:
     return reduce(operator.__or__, imgs)
 
@@ -87,3 +100,42 @@ def make_img_grid(
         grid = img_join_vertical(grid)
 
     return grid
+
+
+# # TODO: Test this
+# def make_img_grid(
+#     imgs: ImageCollection,
+#     ncol=3,
+#     size_WH: Optional[Tuple[int, int]] = (640, 384),
+#     pad=True,
+#     verbose=False
+# ) -> Image.Image:
+#     # fmt: off
+#     if size_WH:
+#         if pad: imgs = [img.resize_pad(size_WH) for img in imgs]
+#         else:   imgs = [img.resize(size_WH) for img in imgs]
+#     else:
+#         pass
+#     # fmt: on
+
+#     rows = assemble_rows(imgs, ncol)
+    
+#     if verbose:
+#         from rich.progress import Progress
+#         with Progress() as prog:
+#             n_tasks = len(rows) + 1
+#             task_concat = prog.add_task("Concatenating Rows...", total=n_tasks)
+
+#             while not prog.finished:
+#                 for row in rows:
+#                     grid += [img_join_horizontal(row)]
+#                     prog.update(task_concat, advance=1)
+
+#                 grid = img_join_vertical(grid)
+#                 prog.update(task_concat, advance=1)
+
+#     else:
+#         grid = [img_join_horizontal(row) for row in rows]
+#         grid = img_join_vertical(grid)
+    
+#     return grid