+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.7.10 -- 26 Jun 2024
 * Bugfixes
 * `image_to_base64` util to convert PIL -> base64 strings
 * UMAP tools: `upyog.ml.umap`